refactor: unify mock file placement for ITabsGateway
- Update import paths in 3 test files to use shared mock
- Delete duplicate createMockTabsGateway.ts from interactor mocks
- Update user-story-008 README with completed tasks

Resolves user-story-008: モックファイル配置規約への統一
