+++ docs/user-stories/user-story-008/README.md
@@ -10,18 +10,24 @@
 
 ## 対象ファイル
 
-以下の重複モックが確認されている:
+以下の重複モックが確認された:
 
-| 重複モック | 正しい共有モック |
-|-----------|-----------------|
-| `tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/mocks/createMockTabsGateway.ts` | `tests/frameworks-and-drivers/browser/ChromeTabsGateway/createMockTabsGateway.ts` |
-
-その他の対象ファイルは別途調査が必要。
+| 重複モック | 正しい共有モック | ステータス |
+|-----------|-----------------|-----------|
+| `tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/mocks/createMockTabsGateway.ts` | `tests/frameworks-and-drivers/browser/ChromeTabsGateway/mocks/createMockTabsGateway.ts` | ✅ 解消済み |
 
 ## タスク
 
-（別途策定）
+- [x] 重複モックファイルの調査・特定
+- [x] 影響を受けるテストファイルの特定
+- [x] テストファイルのインポートパスを正しい共有モックに修正
+  - `normal-cases.test.ts`
+  - `error-cases.test.ts`
+  - `partial-success-cases.test.ts`
+- [x] 重複モックファイルの削除
 
 ## 受け入れ条件
 
-（別途策定）
+- [x] 同一インターフェースのモックが複数箇所に存在しないこと
+- [x] すべてのテストが正しい共有モックを参照していること
+- [x] 重複モックファイルが削除されていること
